requirements.txt
qrcode[pil]==7.4.2
reportlab==4.1.0
Pillow==10.2.0
spotipy==2.23.0

